svg: Update assets for Oreo
Updates (almost) all SVGs to match the new Oreo update.

There are several new faces, people, and activities,
ZWJ sequences, articles of clothing, foods, animals,
and some new objects as well in this release.

For more details, see:
https://emojipedia.org/emoji-5.0/
https://emojipedia.org/google/android-8.0/

(There is still an issue with how SVG flags are dealt with in the
Googlei18n/Noto-Emoji repo, so most flags are still missing
from this repo as of this commit.)

(We do finally have the rainbow flag sequence, though.)