misc: Various fixes and corrections
* Changes version numbers from 1.0 or 1.3 to -> 1.1,

* Changes text mentioning Twemoji or EmojiOne
  to mention Noto Color Emoji instead,

* Minor tweaks/style changes to comment lines or to the README.

Some of this is necessary after cherry-picking commits from
https://github.com/eosrei/emojione-color-font,
but the rest is to address some inaccuracies/style issues
in various files.

(I'm re-writing the commit history,
with an eye toward cherry-picking upstream code where possible,
to preserve attribution to the original authors of the work.

I want my work as separate commits, so as to make it clear/distinct
which work was done by whom. This is possible
now that I have tracked down the original commits.)