lollipop: Create folder for Lollipop builds
Separate folders allows me to keep all editions of each emoji
in the same Git branch ...

which makes it easier to fix old glyphs, as graphical issues
are noticed. This also makes it way easier to manage the hybrid fonts
without so much branch management; i.e. with less cherry-picking
or duplicate commits across different branches.

(All bug fixes should be able to land at the end of the single
active development branch.)

The plan is to use only symbolic links to collect the SVGs
in hybrid folders, and keep originals in the folder for their
respective Android releases.