readme: Replace file with symlink to hybrid README
Replaces the Nougat README.md with (a link to) the hybrid one.

Otherwise, users would be reading the Nougat README.md
every time they visit the hybrid-Nougat-Lollipop branch.

(Which would be kind of weird.)

Going to have to maintain this commit as a patch to the main ("trunk")
branch. Not too big a deal, I suppose. Or I can just drop this commit,
Since the folders are labeled properly...